impr: Rename ryot_bevy to bevy_compass
