+++ ryot_compass/src/bevy_compass/content.rs
