+++ ryot_compass/src/bevy_compass/palette.rs
